azurefunction/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues
azure-storage-blob==12.7.1
azure-functions
aiohttp
pandas
azure-identity
azure-keyvault